refactor: split InlineEdit into typed components
Extract the monolithic InlineEdit component into separate typed
components (InlineText, InlineTextarea, InlineSelect, InlineDatetime)
in a shared components/inline-edit/ directory for better type safety
and reusability.
